# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Entregas",
        "version": "1.0"
    },
    "paths": {
        "/api/componentes": {
            "get": {
                "summary": "Lista todos los componentes",
                "operationId": "15639ece56895b533b1455f180efa545",
                "responses": {
                    "200": {
                        "description": "Lista de componentes"
                    }
                }
            },
            "post": {
                "summary": "Crear un nuevo componente",
                "operationId": "b76347cf9fca4c6ed2378541feded07d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "activo"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "activo": {
                                        "type": "boolean"
                                    },
                                    "imagen": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Componente creado"
                    }
                }
            }
        },
        "/api/componentes/{id}": {
            "get": {
                "summary": "Obtener un componente por ID",
                "operationId": "2fdbece002af2a1b5dccccefc185764b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del componente"
                    }
                }
            },
            "put": {
                "summary": "Actualizar un componente",
                "operationId": "08d34f5236c8b99daed491da8b45133b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "activo": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Componente actualizado"
                    }
                }
            },
            "delete": {
                "summary": "Desactivar un componente",
                "operationId": "20751ed318ec472eca3df038fa165714",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Componente desactivado"
                    }
                }
            }
        },
        "/api/entregas": {
            "get": {
                "summary": "Lista todas las entregas",
                "operationId": "370cd5e04b53833fef92c4c6cda0a82f",
                "responses": {
                    "200": {
                        "description": "Lista de entregas"
                    }
                }
            },
            "post": {
                "summary": "Crear una nueva entrega",
                "operationId": "c10f00ff8aeaf7865d061284613210f1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre_entrega",
                                    "nombre_negocio",
                                    "fecha",
                                    "tipo_boton"
                                ],
                                "properties": {
                                    "nombre_entrega": {
                                        "type": "string"
                                    },
                                    "nombre_negocio": {
                                        "type": "string"
                                    },
                                    "fecha": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "tipo_boton": {
                                        "type": "string",
                                        "enum": [
                                            "wifi",
                                            "ethernet"
                                        ]
                                    },
                                    "observaciones": {
                                        "type": "string"
                                    },
                                    "componentes": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Entrega creada"
                    }
                }
            }
        },
        "/api/entregas/{id}": {
            "get": {
                "summary": "Obtener una entrega por ID",
                "operationId": "1eca0f5594b417b110e6fc58a3313b37",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la entrega"
                    }
                }
            },
            "put": {
                "summary": "Actualizar una entrega",
                "operationId": "2b7ed71ecfaa6b7dae04340c9dfcc2f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre_entrega": {
                                        "type": "string"
                                    },
                                    "nombre_negocio": {
                                        "type": "string"
                                    },
                                    "fecha": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "tipo_boton": {
                                        "type": "string",
                                        "enum": [
                                            "wifi",
                                            "ethernet"
                                        ]
                                    },
                                    "observaciones": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Entrega actualizada"
                    }
                }
            },
            "delete": {
                "summary": "Eliminar una entrega",
                "operationId": "c2a3231b2671cc57d92f8622147243c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Entrega eliminada"
                    }
                }
            }
        }
    }
}